refactor: life is too short to add the same suffix to all files in the folder
