+++ src/commands/confe/UserAssociate.ts
