+++ src/commands/confe/UserVerify.ts
