+++ src/commands/fun/Arrest.ts
