+++ src/commands/misc/Ping.ts
